requirements.txt
PyPDF2
streamlit
langchain
langchain-google-genai
google-generativeai
faiss-cpu
langchain-community